% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{cliFormat}
\alias{cliFormat}
\title{cliFormat}
\usage{
cliFormat(..., .envir = parent.frame())
}
\arguments{
\item{...}{Characters to format}

\item{.envir}{Environment in which to evaluate the expressions}
}
\value{
A formatted character string
}
\description{
Format text into cli inline format
Allows evaluation of expressions within the text before submitting to logs
}
