% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{cliFromLevel}
\alias{cliFromLevel}
\title{cliFromLevel}
\usage{
cliFromLevel(logLevel)
}
\arguments{
\item{logLevel}{Log level as a character string (DEBUG, INFO, SUCCESS, WARN, ERROR, FATAL)}
}
\value{
A function
}
\description{
Get appropriate cli function based on log level
}
\keyword{internal}
