% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osp_print.R
\name{.formatValue}
\alias{.formatValue}
\title{Format a value for display}
\usage{
.formatValue(value)
}
\arguments{
\item{value}{The value to format}
}
\value{
A string representation of the value
}
\description{
Converts a value into a string representation that properly shows its type
}
\keyword{internal}
