% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osp_print.R
\name{.is_empty_value}
\alias{.is_empty_value}
\title{Helper function to check if a value is considered "empty"}
\usage{
.is_empty_value(val)
}
\arguments{
\item{val}{The value to check}
}
\value{
TRUE if the value is considered empty, FALSE otherwise
}
\description{
Determines if a value is empty (NULL, NA, empty string, empty vector, or empty list)
}
\keyword{internal}
