% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum.R
\name{enumRemove}
\alias{enumRemove}
\title{Remove an entry from the enum.}
\usage{
enumRemove(keys, enum)
}
\arguments{
\item{keys}{Key(s) of entries to be removed from the enum}

\item{enum}{Enum from which the entries to be removed
\strong{WARNING}: the original object is not modified!}
}
\value{
Enum without the removed entries.
}
\description{
Remove an entry from the enum.
}
\examples{
Symbol <- enum(c(Diamond = 1, Triangle = 2, Circle = 2))
# either by key
enumRemove("Diamond", Symbol)
# or by positions
enumRemove(2L, Symbol)
}
