% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum.R
\name{enumValues}
\alias{enumValues}
\title{Get the values stored in an enum}
\usage{
enumValues(enum)
}
\arguments{
\item{enum}{\code{enum} containing the values}
}
\value{
List of values stored in the \code{enum}.
}
\description{
Get the values stored in an enum
}
\examples{
Symbol <- enum(c(Diamond = 1, Triangle = 2, Circle = 2))
enumValues(Symbol)
}
\seealso{
Other enumeration-helpers: 
\code{\link{enum}()},
\code{\link{enumGetKey}()},
\code{\link{enumGetValue}()},
\code{\link{enumHasKey}()},
\code{\link{enumKeys}()},
\code{\link{enumPut}()},
\code{\link{enumRemove}()}
}
\concept{enumeration-helpers}
