% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{foldSafe}
\alias{foldSafe}
\title{Safe fold calculation}
\usage{
foldSafe(x, y, epsilon = ospsuiteUtilsEnv$LOG_SAFE_EPSILON)
}
\arguments{
\item{x}{A numeric or a vector of numerics.}

\item{y}{A numeric or a vector of numerics.}

\item{epsilon}{A very small number which is considered as threshold below which
all values are treated as \code{epsilon}. Allows computation of fold changes for values close to 0.
Default value is \code{getOSPSuiteUtilsSetting("LOG_SAFE_EPSILON")}.}
}
\value{
A vector with \code{x / y}
}
\description{
Calculates \code{x / y} while substituting values below \code{epsilon} (for x and y) by \code{epsilon}.
\code{x} and \code{y} must be of the same length
}
\examples{
inputX <- c(NA, 1, 5, 0, -1, defEpsilon)
inputY <- c(1, -1, NA, 0, -1, defEpsilon)
folds <- foldSafe(inputX, inputY)
}
