% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osp_print.R
\name{format_value}
\alias{format_value}
\title{Format a value for display}
\usage{
format_value(value)
}
\arguments{
\item{value}{The value to format}
}
\value{
A string representation of the value
}
\description{
Converts a value into a string representation that properly shows its type
}
\keyword{internal}
