% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-logging.R
\name{getElapsedTime}
\alias{getElapsedTime}
\title{getElapsedTime}
\usage{
getElapsedTime(tic, unit = "min")
}
\arguments{
\item{tic}{Start time}

\item{unit}{display unit of elapsed time}
}
\value{
Character displaying elapsed time in \code{unit}
}
\description{
Get elapsed time between \code{tic} trigger and now
}
\examples{
t0 <- tic()
Sys.sleep(2)
# Get elapsed time in seconds
getElapsedTime(t0, "s")
# Get elapsed time in minutes
getElapsedTime(t0, "min")
}
