% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ospsuite.utils-env.R
\name{getOSPSuiteUtilsSetting}
\alias{getOSPSuiteUtilsSetting}
\title{Get the value of a global ospsuite-R setting.}
\usage{
getOSPSuiteUtilsSetting(settingName)
}
\arguments{
\item{settingName}{String name of the setting}
}
\value{
Value of the setting stored in ospsuiteEnv. If the setting does not exist, an error is thrown.
}
\description{
Get the value of a global ospsuite-R setting.
}
\examples{
getOSPSuiteUtilsSetting("suiteName")
}
