% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error-checks.R
\name{hasEmptyString}
\alias{hasEmptyString}
\title{Check if a string or any of the vector of strings is empty}
\usage{
hasEmptyString(x)
}
\arguments{
\item{x}{A string or a vector of strings.}
}
\value{
\code{TRUE} if any of strings are empty.
}
\description{
Check if a string or any of the vector of strings is empty
}
\details{
If any \code{NA}s are present, they will be removed.
}
\examples{

hasEmptyString(c("x", "y")) # FALSE
hasEmptyString(list("x", "y")) # FALSE
hasEmptyString("   abc   ") # FALSE
hasEmptyString(c("", "y")) # TRUE
hasEmptyString(list("", "y")) # TRUE
hasEmptyString(NA) # TRUE
hasEmptyString(character(0)) # TRUE

}
