% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-emptiness.R
\name{hasEmptyStrings}
\alias{hasEmptyStrings}
\alias{validateHasOnlyNonEmptyStrings}
\title{Validate that no empty string is present}
\usage{
hasEmptyStrings(x)

validateHasOnlyNonEmptyStrings(x)
}
\arguments{
\item{x}{A character string or a vector of character strings.}
}
\value{
\itemize{
\item \code{hasEmptyStrings()} returns \code{TRUE} if any of the strings are empty; \code{FALSE}
otherwise.
\item \code{validateHasOnlyNonEmptyStrings()} produces an error if empty string are
present. It returns \code{NULL} otherwise.
}
}
\description{
Validate that no empty string is present
}
\details{
If any of the following conditions are met, the input string is considered
empty:
\itemize{
\item if any \code{NA}s are present (e.g. \code{x = c("a", "abc", NA)})
\item if string is empty (e.g. \code{x = list("a", "abc", "")})
\item if length is 0 (e.g. \code{x = character()})
}
}
\examples{

hasEmptyStrings(c("x", "y")) # FALSE
hasEmptyStrings(list("x", "y")) # FALSE
hasEmptyStrings("   abc   ") # FALSE
hasEmptyStrings(c("", "y")) # TRUE
hasEmptyStrings(list("", "y")) # TRUE
hasEmptyStrings(NA) # TRUE
hasEmptyStrings(character(0)) # TRUE
hasEmptyStrings(c(NA, "x", "y")) # TRUE

validateHasOnlyNonEmptyStrings(c("x", "y")) # NULL
validateHasOnlyNonEmptyStrings(list("x", "y")) # NULL
validateHasOnlyNonEmptyStrings("   abc   ") # NULL
# validateHasOnlyNonEmptyStrings(c("", "y")) # error
# validateHasOnlyNonEmptyStrings(list("", "y")) # error
# validateHasOnlyNonEmptyStrings(NA) # error
# validateHasOnlyNonEmptyStrings(character(0)) # error
# validateHasOnlyNonEmptyStrings(c(NA, "x", "y")) # error

}
