% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-logging.R
\name{highlight}
\alias{highlight}
\title{highlight}
\usage{
highlight(text)
}
\arguments{
\item{text}{Text to highlight}
}
\description{
Highlight text in console
}
\examples{
cat(paste("This is an", highlight("highlight")))
}
