% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-conditional.R
\name{ifEqual}
\alias{ifEqual}
\title{Value conditional on equality}
\usage{
ifEqual(x, y, outputIfEqual, outputIfNotEqual = NULL)
}
\arguments{
\item{x}{argument 1}

\item{y}{argument 2}

\item{outputIfEqual}{argument 3}

\item{outputIfNotEqual}{argument 4}
}
\description{
Short-key checking if arguments 1 and 2 are equal,
output argument 3 if equal, or output argument 4 otherwise.
}
\examples{

ifEqual(1, 1, "x", "y") # "x"
ifEqual(1, 2, "x", "y") # "y"

}
