% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-conditional.R
\name{ifIncluded}
\alias{ifIncluded}
\title{Value conditional on inclusion}
\usage{
ifIncluded(x, y, outputIfIncluded, outputIfNotIncluded = NULL)
}
\arguments{
\item{x}{argument 1}

\item{y}{argument 2}

\item{outputIfIncluded}{argument 3}

\item{outputIfNotIncluded}{argument 4}
}
\description{
Shortkey checking if arguments 1 is included in 2,
output argument 3 if included, or output argument 4 otherwise.
}
\examples{

ifIncluded("a", c("a", "b"), 1, 2) # 1
ifIncluded("x", c("a", "b"), 1, 2) # 2

}
