% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-conditional.R
\name{ifNotNull}
\alias{ifNotNull}
\title{Value conditional on \code{NULL}}
\usage{
ifNotNull(condition, outputIfNotNull, outputIfNull = NULL)
}
\arguments{
\item{condition}{argument 1}

\item{outputIfNotNull}{argument 2}

\item{outputIfNull}{argument 3}
}
\value{
\code{outputIfNotNull} if condition is not \code{NULL}, \code{outputIfNull} otherwise.
}
\description{
Short-key checking if argument 1 is not \code{NULL}, output the argument 2 if not
null, or output argument 3 otherwise.

Check if condition is not \code{NULL}, if so output \code{outputIfNotNull},
otherwise, output \code{outputIfNull}.
}
\examples{

ifNotNull(NULL, "x")
ifNotNull(NULL, "x", "y")
ifNotNull(1 < 2, "x", "y")

}
