% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-utf.R
\name{isFileUTF8}
\alias{isFileUTF8}
\title{Assess if a file is UTF-8 encoded.}
\usage{
isFileUTF8(file)
}
\arguments{
\item{file}{A name of the file or full path.}
}
\value{
A logical assessing whether there is non UTF-8 encoded characters in \code{file}
}
\description{
Assess if a file is UTF-8 encoded.
}
\examples{

writeLines(c("Hello, world!"), "utf.txt")
writeLines(c("Hello, world!", "\xb5g/L"), "non-utf.txt")

isFileUTF8("utf.txt") # TRUE
isFileUTF8("non-utf.txt") # FALSE

\dontshow{
unlink("utf.txt")
unlink("non-utf.txt")
}

}
