% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error-checks.R
\name{isIncluded}
\alias{isIncluded}
\title{Check if input is included in a list}
\usage{
isIncluded(values, parentValues)
}
\arguments{
\item{values}{Vector of values}

\item{parentValues}{Vector of values}
}
\value{
\code{TRUE} if the values are inside the parent values.
}
\description{
Check if input is included in a list
}
\examples{
A <- data.frame(
  col1 = c(1, 2, 3),
  col2 = c(4, 5, 6),
  col3 = c(7, 8, 9)
)
isIncluded("col3", names(A))
}
