% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-expected-length.R
\name{isOfLength}
\alias{isOfLength}
\alias{validateIsOfLength}
\title{Check if the provided object has expected length}
\usage{
isOfLength(object, nbElements)

validateIsOfLength(object, nbElements)
}
\arguments{
\item{object}{An object or a list of objects}

\item{nbElements}{number of elements that are supposed in object}
}
\value{
\itemize{
\item \code{isOfLength()} returns \code{TRUE} if the object or all objects inside the list
have \code{nbElements}.
\item For \code{validateIsOfLength()}, if validations are successful, \code{NULL} is
returned. Otherwise, error is signaled.
}
}
\description{
Check if the provided object has expected length
}
\note{
Only the first level of the given list is considered.
}
\examples{

df <- data.frame(x = c(1, 2, 3))

isOfLength(df, 1) # TRUE
isOfLength(df, 3) # FALSE

validateIsOfLength(list(1, 2), 2L) # NULL
# validateIsOfLength(c("3", "4"), 3L) # error
}
