% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-paths.R
\name{isPathAbsolute}
\alias{isPathAbsolute}
\alias{validateIsPathAbsolute}
\alias{validatePathIsAbsolute}
\title{Check if path is absolute}
\usage{
isPathAbsolute(path)

validateIsPathAbsolute(path)

validatePathIsAbsolute(path)
}
\arguments{
\item{path}{A valid file path name.}
}
\value{
\itemize{
\item \code{isPathAbsolute()} returns \code{TRUE} if path is absolute (no wildcard); \code{FALSE} otherwise.
\item \code{validateIsPathAbsolute()} returns \code{NULL} if path is absolute. Otherwise, error is signaled.
}
}
\description{
Relative paths will be detected based on the presence of wildcard
character(*) in the path specification.
}
\examples{
# check if path is absolute
isPathAbsolute("Organism|path") # TRUE
isPathAbsolute("Organism|*path") # FALSE

# validation: no error if path is absolute
validateIsPathAbsolute("Organism|path")

# validation: error otherwise
# validateIsPathAbsolute("Organism|*path")
}
