% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error-checks.R
\name{isSameLength}
\alias{isSameLength}
\title{Check if two objects are of same length}
\usage{
isSameLength(...)
}
\arguments{
\item{...}{Objects to compare.}
}
\description{
Check if two objects are of same length
}
\examples{
isSameLength(mtcars, ToothGrowth)
isSameLength(mtcars, mtcars)
}
