% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-utf.R
\name{isUTF8}
\alias{isUTF8}
\title{Assess if a character vector is UTF-8 encoded.}
\usage{
isUTF8(text)
}
\arguments{
\item{text}{A character vector}
}
\value{
A logical assessing whether there is non UTF-8 encoded characters in \code{text}
}
\description{
Assess if a character vector is UTF-8 encoded.
}
\examples{

isUTF8("Hello, world!") # TRUE
isUTF8("\xb5g/L") # FALSE

}
