% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-logger.R
\name{logDebug}
\alias{logDebug}
\title{logDebug}
\usage{
logDebug(msg)
}
\arguments{
\item{msg}{Character values of message to log that leverages \code{cli} formatting.}
}
\description{
Log debug with time stamp
}
\examples{
# Log debug
logDebug("This is a debugging message")

}
\keyword{logging}
