% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-logger.R
\name{logError}
\alias{logError}
\title{logError}
\usage{
logError(msg)
}
\arguments{
\item{msg}{Character values of message to log that leverages \code{cli} formatting.}
}
\description{
Log error with time stamp
}
\examples{
\dontshow{
# Initialize logger settings
setLogFolder()
}
# Log error
logError(cliFormat("This is an {.strong error} message"))

# Log error with indications
logError(cliFormat(
  "This is an {.strong error} message",
  "Check these {.val values} or this {.fn function}"
))

}
\keyword{logging}
