% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{logSafe}
\alias{logSafe}
\title{Computes logarithm of a number or of a vector of numbers and handles zeros while
substituting all values below \code{epsilon} by \code{epsilon}.}
\usage{
logSafe(x, base = exp(1), epsilon = ospsuiteUtilsEnv$LOG_SAFE_EPSILON)
}
\arguments{
\item{x}{A numeric or a vector of numerics.}

\item{base}{a positive or complex number: the base with respect to which logarithms are computed. Defaults to e = exp(1).}

\item{epsilon}{A very small number which is considered as threshold below which
all values are treated as \code{epsilon}. Allows computation of \code{log} close to 0.
Default value is \code{getOSPSuiteUtilsSetting("LOG_SAFE_EPSILON")}.}
}
\value{
\code{log(x, base = base)} for \code{x > epsilon}, or \code{log(epsilon, base = base)},
or \code{NA_real_} for \code{NA} elements.
}
\description{
Computes logarithm of a number or of a vector of numbers and handles zeros while
substituting all values below \code{epsilon} by \code{epsilon}.
}
\examples{
inputVector <- c(NA, 1, 5, 0, -1)
logSafe(inputVector)
}
