% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-logger.R
\name{logWarning}
\alias{logWarning}
\title{logWarning}
\usage{
logWarning(msg)
}
\arguments{
\item{msg}{Character values of message to log that leverages \code{cli} formatting.}
}
\description{
Log warning with time stamp
}
\examples{
# Log warning
logWarning(cliFormat("This is a {.strong warning} message"))

}
\keyword{logging}
