% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\docType{data}
\name{messages}
\alias{messages}
\title{Functions to concat strings to messages used e.g. in errors}
\format{
An object of class \code{list} of length 30.
}
\usage{
messages
}
\description{
Functions to concat strings to messages used e.g. in errors
}
\keyword{datasets}
