% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{msgHeader}
\alias{msgHeader}
\title{msgHeader}
\usage{
msgHeader(logLevel)
}
\arguments{
\item{logLevel}{Log level as a character string (DEBUG, INFO, SUCCESS, WARN, ERROR, FATAL)}
}
\value{
A formatted string with the log level and timestamp
}
\description{
Message header displaying log level and time stamp
}
\keyword{internal}
