% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objCount.R
\name{objCount}
\alias{objCount}
\title{Count number of objects}
\usage{
objCount(x)
}
\arguments{
\item{x}{An object (an atomic vector, a list, or instance(s) of a class).}
}
\value{
Integer representing the count of objects.
}
\description{
Count number of objects
}
\details{
If the argument is not a vector, unlike \code{length()}, this function will not
count the number of named bindings in an environment, but only the number of
instances of a class.

For example, \code{length(mtcars)} will return 11, but \code{objCount(mtcars)} will
return 1.
}
\examples{

objCount(c(1, 2, 3)) # 3
objCount(list("a", "b")) # 2
objCount(mtcars) # 1

}
