% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectCount.R
\name{objectCount}
\alias{objectCount}
\title{Count number of objects}
\usage{
objectCount(x)
}
\arguments{
\item{x}{An object (an atomic vector, a list, or instance(s) of a class).}
}
\value{
Integer representing the count of objects.
}
\description{
Count number of objects
}
\details{
If the argument is not a vector, unlike \code{length()}, this function will not
count the number of named bindings in an environment, but only the number of
instances of a class.

For example, \code{length(mtcars)} will return 11, but \code{objectCount(mtcars)} will
return 1.
}
\examples{

objectCount(c(1, 2, 3)) # 3
objectCount(list("a", "b")) # 2
objectCount(mtcars) # 1
objectCount(list(iris, mtcars)) # 2

}
