% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osp_print.R
\name{osp_print_class}
\alias{osp_print_class}
\title{Print an object's class name}
\usage{
osp_print_class(x)
}
\arguments{
\item{x}{An R object}
}
\value{
Invisibly returns the input object
}
\description{
Prints the class name of an object with nice formatting using cli.
}
\examples{
# Print class name of a data frame
osp_print_class(iris)
}
