% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osp_print.R
\name{ospPrintClass}
\alias{ospPrintClass}
\title{Print an object's class name}
\usage{
ospPrintClass(x)
}
\arguments{
\item{x}{An R object}
}
\value{
Invisibly returns the input object
}
\description{
Prints the class name of an object with nice formatting using cli.
}
\examples{
# Print class name of a data frame
ospPrintClass(iris)
}
\seealso{
Other print functions: 
\code{\link{ospPrintHeader}()},
\code{\link{ospPrintItems}()}
}
\concept{print functions}
