% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ospsuite.utils-env.R
\docType{data}
\name{ospsuiteUtilsSettingNames}
\alias{ospsuiteUtilsSettingNames}
\title{Names of the settings stored in ospsuiteEnv. Can be used with \code{getOSPSuiteUtilsSetting()}}
\format{
An object of class \code{list} of length 3.
}
\usage{
ospsuiteUtilsSettingNames
}
\description{
Names of the settings stored in ospsuiteEnv. Can be used with \code{getOSPSuiteUtilsSetting()}
}
\keyword{datasets}
