% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-logger.R
\name{setInfoMasking}
\alias{setInfoMasking}
\title{setInfoMasking}
\usage{
setInfoMasking(patterns)
}
\arguments{
\item{patterns}{Character patterns to identify with \code{grepl()} when masking messages}
}
\description{
Mask info messages
}
\examples{
\dontrun{
# Mask ggplot2 message when line is used with 1 value per group
setInfoMasking("Each group consists of only one observation")
}
}
\keyword{logging}
