% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-logger.R
\name{setLogFolder}
\alias{setLogFolder}
\title{setLogFolder}
\usage{
setLogFolder(logFolder = NULL)
}
\arguments{
\item{logFolder}{Optional folder path to save log file}
}
\description{
Initialize logs and their settings
}
\keyword{logging}
