% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-logger.R
\name{setWarningMasking}
\alias{setWarningMasking}
\title{setWarningMasking}
\usage{
setWarningMasking(patterns)
}
\arguments{
\item{patterns}{Character patterns to identify with \code{grepl()} when masking messages}
}
\description{
Mask warning messages
}
\examples{
\dontrun{
# Mask ggplot2 warning message when missing values are found
setWarningMasking("rows containing missing values")
}
}
\keyword{logging}
