% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-logger.R
\name{toc}
\alias{toc}
\title{toc}
\usage{
toc(tic, unit = "min")
}
\arguments{
\item{tic}{Start time}

\item{unit}{display unit of elapsed time}
}
\value{
Character displaying elapsed time in \code{unit}
}
\description{
Get elapsed time between \code{tic} trigger and now
}
\examples{
t0 <- tic()
Sys.sleep(2)
# Get elapsed time in seconds
toc(t0, "s")
# Get elapsed time in minutes
toc(t0, "min")
}
