% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{toMissingOfType}
\alias{toMissingOfType}
\title{Convert special constants to \code{NA} of desired type}
\usage{
toMissingOfType(x, type)
}
\arguments{
\item{x}{A single element.}

\item{type}{Type of atomic vector to be returned.}
}
\description{
Convert special constants to \code{NA} of desired type
}
\details{
Special constants (\code{NULL}, \code{Inf}, \code{-Inf}, \code{NaN},  \code{NA}) will be converted to
\code{NA} of desired type.

This function is \strong{not} vectorized, and therefore only scalar values should
be entered.
}
\examples{

toMissingOfType(NA, type = "real")
toMissingOfType(NULL, type = "integer")

}
