% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-uniqueness.R
\name{validateHasOnlyDistinctValues}
\alias{validateHasOnlyDistinctValues}
\title{Validate that vector has unique values}
\usage{
validateHasOnlyDistinctValues(values, na.rm = TRUE)
}
\arguments{
\item{values}{An array of values}

\item{na.rm}{Logical to decide if missing values should be removed from the
duplicate checking. Note that duplicate \code{NA} values are flagged if
\code{na.rm=FALSE}.}
}
\value{
\code{NULL} if only unique values present, otherwise produces error.
}
\description{
Validate that vector has unique values
}
\examples{
validateHasOnlyDistinctValues(c("x", "y")) # NULL
# validateHasOnlyDistinctValues(c("x", "y", "x")) # error

}
