% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-utf.R
\name{validateIsFileUTF8}
\alias{validateIsFileUTF8}
\title{Validate if a file is UTF-8 encoded.}
\usage{
validateIsFileUTF8(file)
}
\arguments{
\item{file}{A name of the file or full path.}
}
\value{
If validations are successful, \code{validateIsFileUTF8()} returns \code{NULL}.
Otherwise, error is signaled.
}
\description{
Validate if a file is UTF-8 encoded.
}
\examples{

writeLines(c("Hello, world!"), "utf.txt")
writeLines(c("Hello, world!", "\xb5g/L"), "non-utf.txt")

validateIsFileUTF8("utf.txt") #NULL
\dontrun{validateIsFileUTF8("non-utf.txt")} # Error

\dontshow{unlink("utf.txt"); unlink("non-utf.txt")}

}
