% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-emptiness.R
\name{validateIsNotEmpty}
\alias{validateIsNotEmpty}
\title{Check that the object is not empty}
\usage{
validateIsNotEmpty(object)
}
\arguments{
\item{object}{An object or an atomic vector or a list of objects.}
}
\value{
If validations are successful, \code{NULL} is returned. Otherwise, error is
signaled.
}
\description{
Check that the object is not empty
}
\examples{
# returns `NULL` if of objects are of specified length
validateIsNotEmpty(list(1, 2))

# error otherwise
# validateIsNotEmpty(NULL)
}
