% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-helpers.R
\name{validateIsOfLength}
\alias{validateIsOfLength}
\title{Check if objects have expected length}
\usage{
validateIsOfLength(object, nbElements)
}
\arguments{
\item{object}{An object or a list of objects}

\item{nbElements}{number of elements that are supposed in object}
}
\value{
If validations are successful, \code{NULL} is returned. Otherwise, error is
signaled.
}
\description{
Check if objects have expected length
}
\examples{
# returns `NULL` if of objects are of specified length
validateIsOfLength(list(1, 2), 2L)

# error otherwise
# validateIsOfLength(c("3", "4"), 3L)
}
