% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-helpers.R
\name{validateIsSameLength}
\alias{validateIsSameLength}
\title{Check if all objects are of same length}
\usage{
validateIsSameLength(...)
}
\arguments{
\item{...}{Objects to compare.}
}
\value{
If validations are successful, \code{NULL} is returned. Otherwise, error is
signaled.
}
\description{
Check if all objects are of same length
}
\examples{
# returns `NULL` if of objects are of same length
validateIsSameLength(list(1, 2), c("3", "4"))

# error otherwise
# validateIsSameLength(list(1, 2), c("3", "4"), c(FALSE))
}
