% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum.R
\name{enumKeys}
\alias{enumKeys}
\title{Get all keys of an enum}
\usage{
enumKeys(enum)
}
\arguments{
\item{enum}{\code{enum} containing the keys.}
}
\value{
List of \code{key} names.
}
\description{
Get all keys of an enum
}
\examples{
Symbol <- enum(c(Diamond = 1, Triangle = 2, Circle = 2))
enumKeys(Symbol)
}
\seealso{
Other enumeration-helpers: 
\code{\link{enumGetKey}()},
\code{\link{enumGetValue}()},
\code{\link{enumHasKey}()},
\code{\link{enumPut}()},
\code{\link{enumRemove}()},
\code{\link{enumValues}()},
\code{\link{enum}()}
}
\concept{enumeration-helpers}
