% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-logger.R
\name{getLogFolder}
\alias{getLogFolder}
\title{getLogFolder}
\usage{
getLogFolder()
}
\description{
Get current log folder where logs are saved
}
\examples{
\dontrun{
# Set/get log folder to a temporary directory
setLogFolder(tempdir())
getLogFolder()

# Set/get logFolder to `NULL`, cancel saving of logs
setLogFolder()
getLogFolder()
}
}
\keyword{logging}
