% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error-checks.R
\name{hasUniqueValues}
\alias{hasUniqueValues}
\title{Remove duplicate values from data}
\usage{
hasUniqueValues(data, na.rm = TRUE)
}
\arguments{
\item{data}{A dataframe.}

\item{na.rm}{Logical to decide if missing values should be removed.}
}
\value{
Logical denoting if there are only unique values in data.
}
\description{
Remove duplicate values from data
}
\examples{
hasUniqueValues(c("x", NA, "y", "x"))
hasUniqueValues(c("x", NA, "y"))
}
