% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error-checks.R
\name{hasUniqueValues}
\alias{hasUniqueValues}
\title{Check that an array of values does not include any duplicate}
\usage{
hasUniqueValues(values, na.rm = TRUE)
}
\arguments{
\item{values}{An array of values}

\item{na.rm}{Logical to decide if missing values should be removed from the duplicate checking.
Note that duplicate \code{NA} values are flagged if \code{na.rm=FALSE}.}
}
\value{
Logical assessing if all values are unique
}
\description{
Check that an array of values does not include any duplicate
}
\examples{
hasUniqueValues(c("x", "y"))
hasUniqueValues(c("x", "y", "x"))
hasUniqueValues(c("x", NA, "y", NA), na.rm = FALSE)
hasUniqueValues(c("x", NA, "y", NA), na.rm = TRUE)
}
