% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-file-extensions.R
\name{isFileExtension}
\alias{isFileExtension}
\alias{validateIsFileExtension}
\title{Validate if the provided path has required extension}
\usage{
isFileExtension(file, extension)

validateIsFileExtension(file, extension)
}
\arguments{
\item{file}{A name of the file or full path.}

\item{extension}{A required extension of the file.}
}
\value{
\code{isFileExtension()} returns \code{TRUE} if the file name (or full path) includes
the extension.

If validations are successful, \code{validateIsFileExtension()} returns \code{NULL}.
Otherwise, error is signaled.
}
\description{
Validate if the provided path has required extension
}
\examples{

isFileExtension("enum.R", "R") # TRUE
isFileExtension("enum.R", "pkml") # FALSE

validateIsFileExtension("enum.R", "R") # NULL
# validateIsFileExtension("enum.R", "pkml") # error

}
