% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error-checks.R
\name{isIncluded}
\alias{isIncluded}
\title{Check if input is included in a list}
\usage{
isIncluded(values, parentValues)
}
\arguments{
\item{values}{Vector of values}

\item{parentValues}{Vector of values}
}
\value{
\code{TRUE} if the values are inside the parent values.
}
\description{
Check if input is included in a list
}
