% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-same-length.R
\name{isSameLength}
\alias{isSameLength}
\alias{validateIsSameLength}
\title{Validate if objects are of same length}
\usage{
isSameLength(...)

validateIsSameLength(...)
}
\arguments{
\item{...}{Objects to compare.}
}
\value{
\itemize{
\item \code{isSameLength()} returns \code{TRUE} if all objects have same lengths.
\item For \code{validateIsSameLength()}, if validations are successful, \code{NULL} is
returned. Otherwise, error is signaled.
}
}
\description{
Validate if objects are of same length
}
\examples{
# compare length of only 2 objects
isSameLength(mtcars, ToothGrowth) # FALSE
isSameLength(cars, BOD) # TRUE

# or more number of objects
isSameLength(c(1, 2), c(TRUE, FALSE), c("x", "y")) # TRUE
isSameLength(list(1, 2), list(TRUE, FALSE), list("x")) # FALSE

# validation
validateIsSameLength(list(1, 2), c("3", "4")) # NULL
# validateIsSameLength(list(1, 2), c("3", "4"), c(FALSE)) # error

}
