% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation-length.R
\name{isSameLength}
\alias{isSameLength}
\title{Check if objects are of same length}
\usage{
isSameLength(...)
}
\arguments{
\item{...}{Objects to compare.}
}
\description{
Check if objects are of same length
}
\examples{
# compare length of only 2 objects
isSameLength(mtcars, ToothGrowth) # FALSE
isSameLength(cars, BOD) # TRUE

# or more number of objects
isSameLength(c(1, 2), c(TRUE, FALSE), c("x", "y")) # TRUE
isSameLength(list(1, 2), list(TRUE, FALSE), list("x")) # FALSE
}
