% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\docType{data}
\name{messages}
\alias{messages}
\title{List of functions and strings used to signal error messages}
\format{
An object of class \code{list} of length 34.
}
\usage{
messages
}
\value{
A string with error message.
}
\description{
Most of these messages will be relevant only in the context of OSP R package
ecosystem.
}
\examples{
# example with string
messages$errorEnumNotAllNames

# example with function
messages$errorPropertyReadOnly("age")

}
\keyword{datasets}
