% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objectCount.R
\name{objectCount}
\alias{objectCount}
\title{Count number of objects}
\usage{
objectCount(x)
}
\arguments{
\item{x}{An object (an atomic vector, a list, or instance(s) of a class).}
}
\value{
Integer representing the count of objects.
}
\description{
Count number of objects
}
\details{
Classes in R can roughly be distinguished as following:
\itemize{
\item Vector-style classes have the property that \code{length()} represents number of
elements (e.g., \code{factor}, \code{list}, etc.).
\item Record-style (or dataframe or scalar) classes, on the other hand, have
complex structures to represent a single thing (e.g., \code{data.frame}, \code{R6},
\code{lm}, etc.).
}

This function counts objects differently depending on the entered class:

If the argument is a vector or a vector-style class, it will return the
output from \code{length()} function. Otherwise, it will returns \code{1}.

For example,
\itemize{
\item \code{length(mtcars)} returns \code{11}, but \code{objectCount(mtcars)} will
return \code{1}, while
\item \code{length(list(1, 2))} returns \code{2}, and \code{objectCount(list(1, 2))} will return
\code{2} as well.
}
}
\examples{

# vectors or vector-style classes
objectCount(c(1, 2, 3)) # 3
objectCount(list("a", "b")) # 2
objectCount(list(iris, mtcars)) # 2

# everything else
objectCount(mtcars) # 1
objectCount(lm(wt ~ mpg, mtcars)) # 1
objectCount(new.env()) # 1

}
