% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{op-null-default}
\alias{op-null-default}
\alias{\%||\%}
\title{Default value for \code{NULL}}
\usage{
x \%||\% y
}
\arguments{
\item{x, y}{If \code{x} is \code{NULL}, will return \code{y}; otherwise returns \code{x}.}
}
\value{
The first object if it is not \code{NULL} otherwise the second object.
}
\description{
Convenience function to avoid testing for \code{NULL}.
}
\examples{
1 \%||\% 2
NULL \%||\% 2

}
