% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osp_print.R
\name{OspPrintClass}
\alias{OspPrintClass}
\title{Print an object's class name}
\usage{
OspPrintClass(x)
}
\arguments{
\item{x}{An R object}
}
\value{
Invisibly returns the input object
}
\description{
Prints the class name of an object with nice formatting using cli.
}
\examples{
# Print class name of a data frame
OspPrintClass(iris)
}
\seealso{
Other print functions: 
\code{\link{OspPrintHeader}()},
\code{\link{OspPrintItems}()}
}
\concept{print functions}
