% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printable.R
\docType{class}
\name{Printable}
\alias{Printable}
\title{Printable}
\description{
Base class that implements some basic properties for printing to console.
}
\examples{
myPrintable <- R6::R6Class(
  "myPrintable",
  inherit = Printable,
  public = list(
    x = NULL,
    y = NULL,
    print = function() {
      private$printClass()
      private$printLine("x", self$x)
      private$printLine("y", self$y)
      invisible(self)
    }
  )
)

x <- myPrintable$new()
x
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Printable-new}{\code{Printable$new()}}
\item \href{#method-Printable-clone}{\code{Printable$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Printable-new"></a>}}
\if{latex}{\out{\hypertarget{method-Printable-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Printable$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Printable-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Printable-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Printable$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
